refactor: rewrite faucet to mjs
