+++ .dockerignore
@@ -3,4 +3,4 @@
 !package.json
 !package-lock.json
 !*.config.js
-!faucet.js
+!faucet.mjs
